// shared/enums.ts
export enum DOCS_EVENT {
    CHANGE_PAGE = "factions:docs:changePage",
    PAGE_UPDATE = "factions:docs:pageUpdate",
    READY = "factions:docs:ready",
    ADD_DOC = "factions:docs:add",
    CREATE_DOC = "factions:docs:create",
    EDIT_DOC = "factions:docs:edit"
}

export enum DOCS_STATE {
    LOCATION = "faction:docs:pages"
}